dev/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.3"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.8.0",
    "firebase-admin>=6.7.0",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
    "qdrant-client>=1.14.2",
]
